Energy Regen works
Energy Regen works with button, based on timeincrement of 5 since last check. Up to Maxenergy(100).
Overflowing of extra energy is also reserved and saved.
CSS Added
Experience is now and Int instead of decimal(both DB and Model)

+++ Data/Context/SQL/GangSQLContext.cs
@@ -78,7 +78,7 @@ namespace Data.Context.SQL
                                     Username = (string)gangMemberReader["username"],
                                     Password = (string)gangMemberReader["password"],
                                     PlayerLevel = (int)gangMemberReader["playerLevel"],
-                                    Experience = (decimal)gangMemberReader["experience"],
+                                    Experience = (int)gangMemberReader["experience"],
                                     SkillPoints = (int)gangMemberReader["skillPoints"],
                                     Money = (decimal)gangMemberReader["money"],
                                     Income = (decimal)gangMemberReader["income"],
@@ -101,13 +101,14 @@ namespace Data.Context.SQL
             }
         }
 
+        // TODO: FIX DEZE Methode als je gangs wilt gebruiken
         public void AddPlayerToGang(int playerId, int gangId)
         {
             try
             {
                 using (SqlConnection conn = _dbConnection.GetConnString())
                 {
-                    // TODO: FIX DEZE STUFF
+                    
                     using (SqlCommand cmd = new SqlCommand("INSERT INTO ", conn))
                     {
